+++ crates/ruma-events/CHANGELOG.md
@@ -9,6 +9,7 @@ Improvements:
 - Don't fail event content parsing on invalid relation
   - We previously already accepted custom or slightly malformed relations
   - Now, even invalid / missing `rel_type` and `event_id` are accepted
+- Implement `From<RoomPowerLevels>` for `ruma_common::push::PushConditionPowerLevelsCtx`
 
 # 0.27.11